requirements.txt
pandas
torch
sentence-transformers
rank_bm25
faiss-cpu
openai
tqdm
python-dotenv
opencc-python-reimplemented
jieba